feat(core): absorb sse into server.py
The sse plugin (/stream/{name} → text/event-stream) was one of the
Tier 0 "can't be a plugin because it needs raw ASGI send" cases
that justified its existence outside the microkernel. But there was
never a reason for it to live in a separate file — the scope of
the handler is small (~75 lines), and keeping it as a plugin required
importing `server` inside `sse.py` just to reach `_valid_name`,
`_disk_name`, `DATA`, and `conn()`.

Inlined as `_core_sse_handle`. Registered via `register_plugin_routes`
alongside `/plugins/propose` and `/plugins/approve`. Auth level `none`
(browsers can't send Authorization on EventSource — this is why sse
was structured as a plugin in the first place but registered with
AUTH="none"; that part transfers).

Heartbeat cadence (25 polls × 20ms = 500ms) preserved — matches the
earlier tune(sse) commit be41146 that took it from 3s to 0.5s. That
tuning was motivated by HTTP/2 edge buffers (Cloudflare tunnel, some
nginx configs) that wait for N bytes before flushing.

plugins/available/sse.py deleted. Backup copy at logs/plugins-backup/.

Part of the microkernel 2C cut. Tests will go red until the
plugins/available/ teardown commits finish and tests are adjusted
— Ranger approved red tests during the cut window.
